move Platform.cpp to common/, seems to fit in a little better there
