+++ src/common/Platform.cpp
