+++ src/common/Platform.h
