+++ src/common/meson.build
@@ -1,4 +1,5 @@
 common_files += files(
+	'Platform.cpp',
 	'String.cpp',
 	'tpt-rand.cpp',
 )
