+++ src/gui/interface/Textbox.cpp
@@ -1,17 +1,17 @@
-#include "gui/interface/Textbox.h"
+#include "Textbox.h"
 
 #include "Config.h"
-#include "Platform.h"
 #include "Format.h"
 #include "PowderToy.h"
 
-#include "graphics/Graphics.h"
+#include "common/Platform.h"
 #include "graphics/FontReader.h"
+#include "graphics/Graphics.h"
 
-#include "gui/interface/Point.h"
+#include "gui/interface/Engine.h"
 #include "gui/interface/Keys.h"
 #include "gui/interface/Mouse.h"
-#include "gui/interface/Engine.h"
+#include "gui/interface/Point.h"
 
 #include "ContextMenu.h"
