+++ src/gui/preview/PreviewController.cpp
@@ -1,21 +1,19 @@
 #include "PreviewController.h"
 
-#include "PreviewView.h"
+#include "Config.h"
+#include "Controller.h"
 #include "PreviewModel.h"
+#include "PreviewModelException.h"
+#include "PreviewView.h"
 
 #include "client/Client.h"
 #include "client/SaveInfo.h"
+#include "common/Platform.h"
 
-#include "PreviewModelException.h"
-
-#include "gui/dialogues/InformationMessage.h"
 #include "gui/dialogues/ErrorMessage.h"
+#include "gui/dialogues/InformationMessage.h"
 #include "gui/login/LoginController.h"
 #include "gui/login/LoginView.h"
-
-#include "Controller.h"
-#include "Platform.h"
-#include "Config.h"
 
 PreviewController::PreviewController(int saveID, int saveDate, bool instant, std::function<void ()> onDone_):
 	saveId(saveID),
