+++ src/gui/profile/ProfileActivity.cpp
@@ -1,16 +1,15 @@
 #include "ProfileActivity.h"
 
-#include "gui/interface/Button.h"
-#include "gui/interface/Textbox.h"
-#include "gui/interface/Label.h"
-#include "gui/interface/AvatarButton.h"
-#include "gui/interface/ScrollPanel.h"
-#include "gui/dialogues/ErrorMessage.h"
+#include "client/Client.h"
+#include "common/Platform.h"
 #include "gui/Style.h"
 
-#include "client/Client.h"
-
-#include "Platform.h"
+#include "gui/interface/AvatarButton.h"
+#include "gui/interface/Button.h"
+#include "gui/dialogues/ErrorMessage.h"
+#include "gui/interface/Label.h"
+#include "gui/interface/ScrollPanel.h"
+#include "gui/interface/Textbox.h"
 
 ProfileActivity::ProfileActivity(ByteString username) :
 	WindowActivity(ui::Point(-1, -1), ui::Point(236, 300)),
