+++ src/gui/save/LocalSaveActivity.cpp
@@ -1,14 +1,12 @@
 #include "LocalSaveActivity.h"
 
 #include "images.h"
-#include "Platform.h"
 
 #include "client/Client.h"
 #include "client/GameSave.h"
 #include "client/ThumbnailRendererTask.h"
-
+#include "common/Platform.h"
 #include "graphics/Graphics.h"
-
 #include "gui/Style.h"
 
 #include "gui/dialogues/ConfirmPrompt.h"
